implementation/backend/pyproject.toml
[project]
name = "chatdf-backend"
version = "0.1.0"
description = "ChatDF backend - FastAPI application for conversational data analysis"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "polars>=1.0.0",
    "google-genai>=1.0.0",
    "authlib>=1.3.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.34",
    "httpx>=0.27",
    "freezegun>=1.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (isolated, mocked dependencies)",
    "integration: Integration tests (real DB, real services)",
    "slow: Tests that take >1s (worker pool, timeouts)",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
